refactor: move cli to handler
