+++ fakesnake/__init__.py
@@ -1,0 +1,2 @@
+from handler import *
+from setting import *
